pyproject.toml
[tool.poetry]
name = "t3"
version = "0.1.0"
description = ""
authors = ["Faisal Albasu <albasfaisal@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyright = "^1.1.385"
streamlit = "^1.39.0"
supabase = "^2.9.0"
python-dotenv = "^1.0.1"
pypdf = "^5.0.1"
textract = "^1.6.5"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonPath = "/opt/homebrew/Caskroom/miniconda/base/envs/dap/bin"
venvPath = "/opt/homebrew/Caskroom/miniconda/base/envs/"
venv = "dap"
typeCheckingMode = "off"
